+++ CHANGELOG.md
@@ -13,12 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `MapFile.parseMapContents`/`MapFile::parse_map_contents`
     - Parses the map contents passed as the argument, without requiring the map
       being on an actual file.
-    - The map format will be guessed on the contents. Currently only the GNU ld
-      map format is recognized.
+    - The map format will be guessed on the contents. Currently both the GNU ld
+      and clang ld.lld map formats are recognized.
   - `MapFile.parseMapContentsGnu`/`MapFile::parse_map_contents_gnu`
     - Parses the map contents passed as the argument, without requiring the map
       being on an actual file.
     - This function only parses the GNU ld map format.
+  - `MapFile.parseMapContentsLld`/`MapFile::parse_map_contents_lld`
+    - Parses the map contents passed as the argument, without requiring the map
+      being on an actual file.
+    - This function only parses the clang ld.lld map format.
 
 ### Changed